Remove some "lazy S" occurrences.
A "lazy S" is a usage of the `S` function in the writing of a unit test, purely as a convenience
for the writer of the test. In other words, this is not a necessary, internal usage of `S` in SymPy;
this was a way to make it easier to write something down.